requirements.txt
django
pythondns